Move the previous `lib.rs` code in `main.rs` to `callbacks.rs` for a (hopefully) cleaner diff.
